+++ config/enquire.php
@@ -10,6 +10,7 @@ return [
         'radio_buttons' => 'enquire::partials.radios',
         'text'          => 'enquire::partials.text',
         'textarea'      => 'enquire::partials.textarea',
+        'current-url'   => 'enquire::partials.current-url',
     ],
 
     'fields_with_options' => [