+++ resources/views/partials/phone.blade.php
@@ -1,5 +1,5 @@
 {{ Form::label($field->name, $field->label.($field->required ? ' *' : ''), ['class' => 'sr-only']) }}
-{{ Form::number($field->name, null, array_filter([
+{{ Form::tel($field->name, null, array_filter([
     'class' => 'form__input',
     'placeholder' => $field->placeholder,
     'required' => $field->required