+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 2.4.0 (Future)
 - Add jam field for selecting forms on entities
 - Add radio field type
+- Separate form partial for manual, non view composer usage
 
 ## Version 2.3.6 (2017-02-08)
 - Fix validation for files